pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reactor-agent"
version = "0.1.0"
description = "A ReAct (Reasoning + Acting) agent framework for AI-powered reasoning and tool execution"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Jacques Lemieux", email = "jacques.lemieux@gmail.com"}
]
keywords = ["ai", "agent", "reasoning", "react", "tools", "openai", "tavily"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.8"
dependencies = [
    "openai>=1.98.0",
    "tavily-python>=0.7.10",
    "pydantic>=2.11.7",
    "httpx>=0.28.1",
    "requests>=2.32.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.0",
    "flake8>=7.3.0",
    "build>=1.2.2",
]

[project.urls]
Homepage = "https://github.com/jalemieux/reactor-agent"
Repository = "https://github.com/jalemieux/reactor-agent"
"Bug Tracker" = "https://github.com/jalemieux/reactor-agent/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["reactor*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"] 